[FIX] 테스트코드 수정 (Auth, Member) 분리
